Move all classes associated with a specific pack into their own folders
