+++ src/main/java/thePackmaster/cards/dimensiongatepack/Inferno.java
@@ -5,7 +5,7 @@ import com.megacrit.cardcrawl.cards.AbstractCard;
 import com.megacrit.cardcrawl.cards.DamageInfo;
 import com.megacrit.cardcrawl.characters.AbstractPlayer;
 import com.megacrit.cardcrawl.monsters.AbstractMonster;
-import thePackmaster.actions.SelfDamageAction;
+import thePackmaster.actions.dimensiongatepack.SelfDamageAction;
 
 import static thePackmaster.SpireAnniversary5Mod.makeID;
