+++ src/main/java/thePackmaster/cards/downfallpack/GhostflameStrike.java
@@ -5,7 +5,7 @@ import com.megacrit.cardcrawl.actions.defect.ChannelAction;
 import com.megacrit.cardcrawl.characters.AbstractPlayer;
 import com.megacrit.cardcrawl.monsters.AbstractMonster;
 import thePackmaster.cards.AbstractPackmasterCard;
-import thePackmaster.orbs.Ghostflame;
+import thePackmaster.orbs.downfallpack.Ghostflame;
 
 import static thePackmaster.SpireAnniversary5Mod.makeID;
 import static thePackmaster.util.Wiz.*;
