+++ src/main/java/thePackmaster/cards/downfallpack/ShapersBlessing.java
@@ -5,7 +5,7 @@ import com.megacrit.cardcrawl.characters.AbstractPlayer;
 import com.megacrit.cardcrawl.monsters.AbstractMonster;
 import com.megacrit.cardcrawl.powers.PlatedArmorPower;
 import thePackmaster.cards.AbstractPackmasterCard;
-import thePackmaster.stances.AncientStance;
+import thePackmaster.stances.downfallpack.AncientStance;
 
 import static thePackmaster.SpireAnniversary5Mod.makeID;
 import static thePackmaster.util.Wiz.applyToSelf;
