+++ src/main/java/thePackmaster/vfx/legacypack/ShootAnythingEffect.java
@@ -1,4 +1,4 @@
-package thePackmaster.vfx;
+package thePackmaster.vfx.legacypack;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.Color;
